+++ packages/backend/src/services/CoderService/CoderService.ts
@@ -91,6 +91,7 @@ export class CoderService extends BaseService {
     private static transformChart(
         chart: SavedChartDAO,
         spaceSummary: Pick<SpaceSummary, 'uuid' | 'slug'>[],
+        dashboardSlugs: Record<string, string>,
     ): ChartAsCode {
         const spaceSlug = spaceSummary.find(
             (space) => space.uuid === chart.spaceUuid,
@@ -106,7 +107,9 @@ export class CoderService extends BaseService {
             updatedAt: chart.updatedAt,
             metricQuery: chart.metricQuery,
             chartConfig: chart.chartConfig,
-            dashboardUuid: chart.dashboardUuid,
+            dashboardSlug: chart.dashboardUuid
+                ? dashboardSlugs[chart.dashboardUuid]
+                : undefined,
             slug: chart.slug,
             tableConfig: chart.tableConfig,
 
@@ -183,9 +186,12 @@ export class CoderService extends BaseService {
                 isChartTile(tile) ? [...acc, tile.properties.chartSlug] : acc,
             [],
         );
+
         const charts = await this.savedChartModel.find({
             slugs: chartSlugs,
             projectUuid,
+            excludeChartsSavedInDashboard: false,
+            includeOrphanChartsWithinDashboard: true,
         });
 
         return tiles.map((tile) => {
@@ -193,6 +199,7 @@ export class CoderService extends BaseService {
                 const savedChart = charts.find(
                     (chart) => chart.slug === tile.properties.chartSlug,
                 );
+
                 if (!savedChart) {
                     throw new NotFoundError(
                         `Chart with slug ${tile.properties.chartSlug} not found`,
@@ -225,9 +232,9 @@ export class CoderService extends BaseService {
 
         if (uuids.length > 0) {
             if (type === 'dashboard') {
-                uuidsToSlugs = await this.dashboardModel.getSlugsForUuids(
-                    uuids,
-                );
+                const dashboardSlugs =
+                    await this.dashboardModel.getSlugsForUuids(uuids);
+                uuidsToSlugs = Object.values(dashboardSlugs);
             } else if (type === 'chart') {
                 uuidsToSlugs = await this.savedChartModel.getSlugsForUuids(
                     uuids,
@@ -388,6 +395,7 @@ export class CoderService extends BaseService {
             projectUuid,
             slugs,
             excludeChartsSavedInDashboard: false,
+            includeOrphanChartsWithinDashboard: true,
         });
         const maxResults = this.lightdashConfig.contentAsCode.maxDownloads;
 
@@ -410,10 +418,20 @@ export class CoderService extends BaseService {
         // get all spaces to map  spaceSlug
         const spaceUuids = charts.map((chart) => chart.spaceUuid);
         const spaces = await this.spaceModel.find({ spaceUuids });
+        // get all spaces to map  dashboardSlug
+        const dashboardUuids = charts.reduce<string[]>((acc, chart) => {
+            if (chart.dashboardUuid) {
+                acc.push(chart.dashboardUuid);
+            }
+            return acc;
+        }, []);
+        const dashboards = await this.dashboardModel.getSlugsForUuids(
+            dashboardUuids,
+        );
 
         return {
             charts: charts.map((chart) =>
-                CoderService.transformChart(chart, spaces),
+                CoderService.transformChart(chart, spaces, dashboards),
             ),
             missingIds,
             total: chartSummaries.length,
@@ -446,6 +464,7 @@ export class CoderService extends BaseService {
             slug,
             projectUuid,
             excludeChartsSavedInDashboard: false,
+            includeOrphanChartsWithinDashboard: true,
         });
 
         // If chart does not exist, we can't use promoteService,
@@ -462,17 +481,58 @@ export class CoderService extends BaseService {
                 `Creating chart "${chartAsCode.name}" on project ${projectUuid}`,
             );
 
-            const createChart: CreateSavedChart & {
+            let createChart: CreateSavedChart & {
                 updatedByUser: UpdatedByUser;
                 slug: string;
                 forceSlug: boolean;
-            } = {
-                ...chartAsCode,
-                spaceUuid: space.uuid,
-                dashboardUuid: undefined, // TODO for charts within dashboards, we need to create the dashboard first, use promotion for that
-                updatedByUser: user,
-                forceSlug: true, // do not generate a new unique slug, use the one from chart as code , at this point we know it is going to be unique on this project
             };
+
+            if (chartAsCode.dashboardSlug) {
+                const [dashboard] = await this.dashboardModel.find({
+                    projectUuid,
+                    slug: chartAsCode.dashboardSlug,
+                });
+
+                let dashboardUuid: string = dashboard?.uuid;
+                if (!dashboard) {
+                    // Charts within dashboards need a dashboard first,
+                    // so we will create a placeholder dashboard for this
+                    // which we can update later
+                    console.debug(
+                        'Creating placeholder dashboard for chart within dashboard',
+                        chartAsCode.slug,
+                    );
+                    const newDashboard = await this.dashboardModel.create(
+                        space.uuid,
+                        {
+                            name: friendlyName(chartAsCode.dashboardSlug),
+                            tiles: [],
+                            slug: chartAsCode.dashboardSlug,
+                            forceSlug: true,
+                            tabs: [],
+                        },
+                        user,
+                        projectUuid,
+                    );
+
+                    dashboardUuid = newDashboard.uuid;
+                }
+                createChart = {
+                    ...chartAsCode,
+                    spaceUuid: null,
+                    dashboardUuid,
+                    updatedByUser: user,
+                    forceSlug: true,
+                };
+            } else {
+                createChart = {
+                    ...chartAsCode,
+                    spaceUuid: space.uuid,
+                    dashboardUuid: null,
+                    updatedByUser: user,
+                    forceSlug: true,
+                };
+            }
 
             const newChart = await this.savedChartModel.create(
                 projectUuid,
@@ -517,6 +577,7 @@ export class CoderService extends BaseService {
                 user,
                 projectUuid, // We use the same projectUuid for both promoted and upstream
                 chart.uuid,
+                true, // includeOrphanChartsWithinDashboard
             );
         const updatedChart = {
             ...promotedChart,
@@ -534,7 +595,6 @@ export class CoderService extends BaseService {
             updatedChart,
             upstreamChart,
         );
-
         promotionChanges = await this.promoteService.upsertCharts(
             user,
             promotionChanges,
@@ -676,6 +736,7 @@ export class CoderService extends BaseService {
                 },
                 projectUuid, // We use the same projectUuid for both promoted and upstream
             );
+
         PromoteService.checkPromoteDashboardPermissions(
             user,
             promotedDashboard,
@@ -701,6 +762,7 @@ export class CoderService extends BaseService {
                 user,
                 promotedDashboard,
                 upstreamDashboard,
+                true, // includeOrphanChartsWithinDashboard
             );
 
         // TODO Right now dashboards on promote service always update dashboards
@@ -710,6 +772,7 @@ export class CoderService extends BaseService {
             user,
             promotionChanges,
         );
+
         promotionChanges = await this.promoteService.upsertCharts(
             user,
             promotionChanges,
@@ -724,7 +787,6 @@ export class CoderService extends BaseService {
         console.info(
             `Finished updating dashboard "${dashboard.name}" on project ${projectUuid}: ${promotionChanges.dashboards[0].action}`,
         );
-
         return promotionChanges;
     }
 }