+++ scripts/clean-repo.sh
@@ -1,12 +1,13 @@
-# Remove TypeScript build info files
-find packages -name "tsconfig*.tsbuildinfo" -exec rm {} +
+rm -rf packages/*/tsconfig*.tsbuildinfo
 
 # Remove node_modules directories
-find packages -name "node_modules" -type d -exec rm -rf {} +
 rm -rf node_modules
+rm -rf packages/*/node_modules
 
 # Remove build/dist directories
-find packages -type d \( -name "build" -o -name "dist" \) -exec rm -rf {} +
+rm -rf packages/*/build
+rm -rf packages/*/dist
 
 # Reinstall dependencies
-yarn
+pnpm install
+pnpm build